FIX-6 :sparkles: : AdateController api 문서 작성
